+++ spring5x-mybatis-base/src/main/java/com/zja/service/UserService.java
@@ -26,7 +26,7 @@ public interface UserService {
     int oracleBatchSaveUser(List<UserEntity> userEntities);
 
     //按id查询用户
-    UserEntity queryUserById(Integer id);
+    UserEntity queryUserById(Long id);
     //查询所有用户
     List<UserEntity> queryAllUser();
     //按用户查询用户信息
@@ -41,7 +41,7 @@ public interface UserService {
     int updateUser(UserEntity userEntity);
 
     //删除数据
-    int delUser(Integer id);
+    int delUser(Long id);
 
 
 }
